Rename project to ConsoleFlow
