+++ src/ConsoleFlow/UI/HorizontalBox.cs
@@ -1,9 +1,9 @@
 using System;
 using System.Collections.Generic;
 
-using TerminalFlow.Core;
+using ConsoleFlow.Core;
 
-namespace TerminalFlow
+namespace ConsoleFlow
 {
     /// <summary>
     ///
